Renamed Validator class
