+++ src/main/java/co/arago/hiro/client/rest/AppAPI.java
@@ -9,8 +9,8 @@ import co.arago.hiro.client.util.httpclient.HttpResponseParser;
 
 import java.io.IOException;
 
-import static co.arago.util.validation.RequiredFieldChecks.notBlank;
-import static co.arago.util.validation.RequiredFieldChecks.notEmpty;
+import static co.arago.util.validation.ValueChecks.notBlank;
+import static co.arago.util.validation.ValueChecks.notEmpty;
 
 public class AppAPI extends AuthenticatedAPIHandler {
