+++ src/main/java/co/arago/hiro/client/websocket/ActionWebSocket.java
@@ -24,7 +24,7 @@ import java.io.IOException;
 import java.net.http.WebSocket;
 import java.util.Map;
 
-import static co.arago.util.validation.RequiredFieldChecks.notNull;
+import static co.arago.util.validation.ValueChecks.notNull;
 
 /**
  * The handler for Action WebSocket.
